requirements.txt
python-dotenv
together
flask
flask-cors
opencv-python
